vit_disease/requirements.txt
python-telegram-bot==20.3
transformers==4.44.0
torch
torchvision
pillow
openai
python-dotenv
